GloOrb explosion mechanic is functional.
Changed the entry point for the enemy death so that things can be added before the death happens.

Modified so that the behavior tree starts in begin play on the enemy instead of it starting in different places for different AI.

Modified Behavior Tree tasks to cast to the generic enemy class instead of a concrete one.